pyproject.toml
[tool.poetry]
name = "hobby-project"
version = "0.1.0"
description = ""
authors = ["Shakhlyn <shakhlyn.sh.du@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.12"}
uvicorn = {extras = ["standard"], version = "^0.34.3"}
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.41"
tenacity = "^9.1.2"
psycopg2-binary = "^2.9.10"
py-cord = "^2.6.1"
python-dotenv = "^1.1.1"
apscheduler = "^3.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
